# Speed up the query for determining if a version is the first occurrence within a job