## What does this PR accomplish?

Bug Fix | **Feature** | Documentation

On our large deployment, we were seeing that the first occurrence query was a bottleneck for the database. 

https://github.com/concourse/concourse/blob/edae11afb7b9baa021eb28a66da02b464e9fdf66/atc/db/versions_db.go#L38-L47

There were consistently 20 of these queries being run on the database, which ends up clogging up the database connections. When I ran an `EXPLAIN ANALYZE` on the query using a job that had 3000 builds, I saw that it was taking almost 15 seconds to run

```
Result  (cost=3317.61..3317.62 rows=1 width=1) (actual time=14950.212..14950.213 rows=1 loops=1)
  CTE builds_of_job
    ->  Index Only Scan using builds_ordered_by_job_id_idx on builds  (cost=0.43..3288.58 rows=860 width=4) (actual time=6.476..14916.978 rows=3190 loops=1)
          Index Cond: (job_id = 93253)
          Heap Fetches: 3190
  InitPlan 2 (returns $1)
    ->  Hash Join  (cost=8.59..29.03 rows=1 width=0) (actual time=14950.211..14950.211 rows=1 loops=1)
          Hash Cond: (b.id = i.build_id)
          ->  CTE Scan on builds_of_job b  (cost=0.00..17.20 rows=860 width=4) (actual time=6.478..14922.335 rows=3190 loops=1)
          ->  Hash  (cost=8.58..8.58 rows=1 width=4) (actual time=24.461..24.461 rows=6 loops=1)
                Buckets: 1024  Batches: 1  Memory Usage: 9kB
                ->  Index Scan using build_inputs_resource_versions_idx on build_resource_config_version_inputs i  (cost=0.56..8.58 rows=1 width=4) (actual time=24.452..24.453 rows=6 loops=1)
                      Index Cond: ((resource_id = 31188) AND (version_md5 = 'b5029cead24551c69709428a9543e078'::text))
                      Filter: (name = 'repository'::text)
Planning time: 0.282 ms
Execution time: 14950.289 ms
```

It looks like most of the time that it takes is due to the fetching of all the builds for that job. Knowing this, I can think of two ways to fix this: 
1. Refactor the query to make it not need to fetch all the builds for a job
2. Denormalize the `job_id` column by adding it to the `build_resource_config_version_inputs` table

I think each of these options have its good and bad. 

**For the first option with refactoring the query,** the way that I can think of to do this is to make it narrow down the builds **by version first** and then scope it down to the job id. This way there will be a smaller number of builds to work with, since the number of builds that match the version/resource will **most likely and more commonly** be smaller than the number of builds within a job. I say this with uncertainty because I understand that might not always be the case, for example, there might be some versions that are being used throughout many builds within a pipeline (a git version that has not been updated for 2 years but is being used for every build within several jobs). The following query would be how I imagine it would be written:

```
EXPLAIN ANALYZE WITH builds_of_version AS ( 
	SELECT build_id
	FROM build_resource_config_version_inputs
	WHERE name = 'repository'
	AND version_md5 = 'b5029cead24551c69709428a9543e078' 
	AND resource_id = 31188
) SELECT EXISTS ( 
	SELECT 1 
	FROM builds b
	JOIN builds_of_version v ON b.id = v.build_id
	WHERE job_id = 93253 
);
```

I tried running this option with `EXPLAIN ANALYZE` and this is the results I got.

```
Result  (cost=17.06..17.07 rows=1 width=1) (actual time=30.810..30.811 rows=1 loops=1)
  CTE builds_of_version
    ->  Index Scan using build_inputs_resource_versions_idx on build_resource_config_version_inputs  (cost=0.56..8.58 rows=1 width=4) (actual time=15.400..15.410 rows=5 loops=1)
          Index Cond: ((resource_id = 31188) AND (version_md5 = 'b5029cead24551c69709428a9543e078'::text))
          Filter: (name = 'repository'::text)
  InitPlan 2 (returns $2)
    ->  Nested Loop  (cost=0.43..8.48 rows=1 width=0) (actual time=30.807..30.807 rows=1 loops=1)
          ->  CTE Scan on builds_of_version v  (cost=0.00..0.02 rows=1 width=4) (actual time=15.402..15.416 rows=5 loops=1)
          ->  Index Only Scan using builds_ordered_by_job_id_idx on builds b  (cost=0.43..8.45 rows=1 width=4) (actual time=3.076..3.076 rows=0 loops=5)
                Index Cond: ((job_id = 93253) AND (id = v.build_id))
                Heap Fetches: 1
Planning time: 0.414 ms
Execution time: 30.864 ms
```

Which is a lot faster as it took 30 ms vs 15000 ms. But this is also because I chose a job that had 3000 builds, which means that in this specific case it would be faster to use the refactored query. If, lets say, I chose an example that had a version used throughout many builds then the results might not be so much of an improvement. One thing to note is that if there is a version that is used throughout many builds, most likely it would be used throughout many builds **of the same job** so that means that both the original and refactored query would be slow in that situation. 

**For the second option with denormalizing,** this means adding a `job_id` column to the `build_resource_config_version_inputs` table. We would then be able to just do a simple query without any CTE to find a version of a resource within a job. It would look something like this

```
EXPLAIN ANALYZE SELECT EXISTS ( 
	SELECT 1 
	FROM builds
	WHERE name = 'repository'
	AND version_md5 = 'b5029cead24551c69709428a9543e078' 
	AND resource_id = 31188
	WHERE job_id = 93253 
);
``` 

I am unable to actually run the query without the column being there but I imagine it would be incredibly fast to run. This would remove any slowness due to needing to narrow down the set of builds first using a CTE, whether it is through a job_id or version, and would be directly using an index to fetch that exact row that we are looking for. This is probably the _ideal_ choice, but it comes with the drawbacks of having to add a migration that will need to run for the incredibly large `build_resource_config_version_inputs` table to add the `job_id` into the table. It will also result in duplication of the `job_id` column, as the data already exists within the `builds` table. This is very similar to what already exists within the `successful_build_outputs` table (where it has a `job_id`) and also what is currently being added to the `build_image_resource_caches` table https://github.com/concourse/concourse/pull/5993. So I'm just unsure how far we want to go with this, or if it is actually a good idea to just spread that data throughout tables that we know are going to be really large and will benefit from this denormalization.

## Changes proposed by this PR:

The change introduces a refactor to the first occurrence query that should help improve its speed. Previously, the query would fetch all the builds within a job and then try to find a version within those list of builds that match the version. But that causes a lot of slowness when the job has a looot of builds. (the CTE part of the query is slow)

The fix here is to have the first part of the query (the CTE) to find all the builds that have that version and then try to find a build that matches the job we are looking for. This will hopefully help reduce the number of rows from the CTE, because there are probably less builds that match the version than there are builds that are from the job. But at least in the most common case, I believe that there will be less rows from narrowing down with a version vs narrowing down with a job id.

## Notes to reviewer:

I submitted this PR with the first option but I actually just wanted to start a conversation in order to find what the best decision is going forward. The solution doesn't need to be one of these two options, those are just the ones I thought of at this moment. I am totally open to doing anything else we come up with!

## Release Note
<!--
If needed, you can leave a list of detailed descriptions here which will be
used to generate the release note for the next version of Concourse. The title
of the PR will also be pulled into the release note.

Example:
* Reticulating splines is the new process Concourse uses to create the network
  of lines between jobs.
* Combines many short lines and curves into a network of splines.
-->

* <!-- Release note here. Delete section if not needed. -->

## Contributor Checklist
<!--
Most of the PRs should have the following added to them,
this doesn't apply to all PRs, so it is helpful to tell us what you did.
-->
- [ ] Followed [Code of conduct], [Contributing Guide] & avoided [Anti-patterns]
- [ ] [Signed] all commits
- [ ] Added tests (Unit and/or Integration)
- [ ] Updated [Documentation]
- [ ] Added release note (Optional)

[Code of Conduct]: https://github.com/concourse/concourse/blob/master/CODE_OF_CONDUCT.md
[Contributing Guide]: https://github.com/concourse/concourse/blob/master/CONTRIBUTING.md
[Anti-patterns]: https://github.com/concourse/concourse/wiki/Anti-Patterns
[Signed]: https://help.github.com/en/github/authenticating-to-github/signing-commits
[Documentation]: https://github.com/concourse/docs

## Reviewer Checklist
<!--
This section is intended for the reviewers only, to track review
progress.
-->
- [ ] Code reviewed
- [ ] Tests reviewed
- [ ] Documentation reviewed
- [ ] Release notes reviewed
- [ ] PR acceptance performed
- [ ] New config flags added? Ensure that they are added to the
  [BOSH](https://github.com/concourse/concourse-bosh-release) and
  [Helm](https://github.com/concourse/helm) packaging; otherwise, ignored for
  the [integration
  tests](https://github.com/concourse/ci/tree/master/tasks/scripts/check-distribution-env)
  (for example, if they are Garden configs that are not displayed in the
  `--help` text).