@aoldershaw 2021-01-11
Denormalizing here is a bit spooky on such a huge table - the increased disk space isn't ideal, but more concerning to me is the migration that it requires. If there was a way to denormalize `job_id` here without making the upgrade more risky, I think I'd be for it, especially given the precedent on `successful_build_outputs` and `build_image_resource_caches`.

Without denormalization, I definitely expect this to be faster in the vast majority of cases, but as you note, some versions are used by many builds. e.g. on CI, version 1.2.2 of [`dumb-init`](https://ci.concourse-ci.org/teams/main/pipelines/concourse/resources/dumb-init) has been around since 2018, and has amassed 1797 builds - which is more than the number of builds on the `dev-image` job (~700), so we'd see a performance dip here.

More dramatic examples are [`oxygen-mask`](https://ci.concourse-ci.org/teams/main/pipelines/monitoring-stats/resources/oxygen-mask) and [`fly`](https://ci.concourse-ci.org/teams/main/pipelines/monitoring-stats/resources/fly) in the [`monitoring-stats`](https://ci.concourse-ci.org/teams/main/pipelines/monitoring-stats) pipeline. Each job is currently in the order of ~100,000 builds, and the same resource is being shared between 5 jobs. On CI, version `cf462dd98d3c5f2129b5c86bac835305` of `oxygen-mask` currently has over 400000 entries in `build_resource_config_version_inputs`, for instance.

So with this change, you've greatly optimized the query for jobs that have frequently changing versions as inputs, but long-lived versions may take a performance hit. The thing about these long-lived versions is that they rarely change build-to-build (obviously) - so what if we added a hot path for that case.

* Since we just need to find one previous instance of the version for the job, we could first check the version from the *previous build of the job* - if it's the same version, then we're done
  * If it's the same version, we're done
  * Otherwise, we run the potentially slow query

This would introduce a slight performance decrease for jobs with inputs that change frequently, but with the optimization you've already added, both queries should be quick in this case, so it hopefully wouldn't be too bad. For long-lived versions, we may still have to run the slow query (typically if an old version is pinned), but it should be much more rare.

WDYT?