@clarafu 2021-01-12
I definitely agree, seeing that data about those versions being used by so many builds is just confirmation that it might be super slow for some use cases. Just to get some more data, I ran the modified query using the `oxygen-mask` resource with the `version_md5 ` of `cf462dd98d3c5f2129b5c86bac835305` and this is what I get as the result

```
Result  (cost=66302.35..66302.36 rows=1 width=1) (actual time=69.091..69.091 rows=1 loops=1)
  CTE builds_of_version
    ->  Bitmap Heap Scan on build_resource_config_version_inputs  (cost=1665.49..66096.31 rows=5459 width=8) (actual time=69.062..69.062 rows=1 loops=1)
          Recheck Cond: ((resource_id = 11775) AND (version_md5 = 'cf462dd98d3c5f2129b5c86bac835305'::text))
          Filter: (name = 'oxygen-mask'::text)
          Heap Blocks: exact=1
          ->  Bitmap Index Scan on build_inputs_resource_versions_idx  (cost=0.00..1664.13 rows=35557 width=0) (actual time=57.969..57.969 rows=408109 loops=1)
                Index Cond: ((resource_id = 11775) AND (version_md5 = 'cf462dd98d3c5f2129b5c86bac835305'::text))
  InitPlan 2 (returns $2)
    ->  Nested Loop  (cost=0.43..23233.97 rows=113 width=0) (actual time=69.089..69.089 rows=1 loops=1)
          ->  CTE Scan on builds_of_version v  (cost=0.00..109.18 rows=5459 width=8) (actual time=69.065..69.065 rows=1 loops=1)
          ->  Index Only Scan using builds_ordered_by_job_id_idx on builds b  (cost=0.43..4.23 rows=1 width=8) (actual time=0.017..0.018 rows=1 loops=1)
                Index Cond: ((job_id = 5946) AND (id = v.build_id))
                Heap Fetches: 1
Planning time: 0.451 ms
Execution time: 70.635 ms
```

Which is suuuper slow compared to if I run the same inputs against the original query that takes about 0.1 ms to complete. 

The idea you proposed is great! It would help to speed up the query for versions that have been used for many builds but still keeping the speed of versions that are changing often. I think the only times that it might have some performance drops are uncommon use cases like you mentioned, when versions are messed with.

So I was hoping to see if there can be some optimization that could make the query faster even for versions that are used for many builds because I think (of course) ideally there are no slow queries even during uncommon cases. I think if we add this index `create index someindex on build_resource_config_version_inputs (resource_id, version_md5, name, build_id);` it will help speed up the possibly slow query. If I run the exact same inputs and modified query using this index, it runs in 0.2 ms

```
Result  (cost=8683.46..8683.47 rows=1 width=1) (actual time=0.187..0.187 rows=1 loops=1)
  CTE builds_of_version
    ->  Index Only Scan using deleteme3 on build_resource_config_version_inputs  (cost=0.56..8477.38 rows=5460 width=8) (actual time=0.165..0.165 rows=1 loops=1)
          Index Cond: ((resource_id = 11775) AND (version_md5 = 'cf462dd98d3c5f2129b5c86bac835305'::text) AND (name = 'oxygen-mask'::text))
          Heap Fetches: 1
  InitPlan 2 (returns $2)
    ->  Nested Loop  (cost=0.43..23238.45 rows=113 width=0) (actual time=0.187..0.187 rows=1 loops=1)
          ->  CTE Scan on builds_of_version v  (cost=0.00..109.20 rows=5460 width=8) (actual time=0.168..0.168 rows=1 loops=1)
          ->  Index Only Scan using builds_ordered_by_job_id_idx on builds b  (cost=0.43..4.23 rows=1 width=8) (actual time=0.013..0.013 rows=1 loops=1)
                Index Cond: ((job_id = 5946) AND (id = v.build_id))
                Heap Fetches: 0
Planning time: 0.403 ms
Execution time: 0.228 ms
```

If we do add that index, we can remove the `build_inputs_resource_versions_idx` index because it will be a covered by this new index.