@aoldershaw 2021-01-19
Interesting! Just adding an index seems like a better approach than I suggested given that it works in the general case.

Sounds like a pretty big index though - it's basically a complete clone of the table! Do we have an idea of how costly that migration would be? Or the additional storage requirements? Mostly just playing devil's advocate, I think your approach makes a lot of sense!